requirements.txt
langchain
langchain-anthropic
anthropic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytest
flask
